requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
sentence-transformers==2.2.2
numpy==1.25.0
faiss-cpu==1.7.4       
google-genai==0.1.0    
pydantic==1.10.9
typing-extensions==4.6.3